Move contact migrate code from message.js to contact.js
